requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
Django==5.1.2
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-rest-framework==0.1.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.10
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
